Updates test to validate subscription and csv (#4)
* Updates test to validate subscription and csv
* Removes platform navigator dependency

+++ .github/scripts/validate-deploy.sh
@@ -2,6 +2,7 @@
 
 GIT_REPO=$(cat git_repo)
 GIT_TOKEN=$(cat git_token)
+BIN_DIR=$(cat .bin_dir)
 
 export KUBECONFIG=$(cat .kubeconfig)
 NAMESPACE="openshift-operators"
@@ -10,7 +11,7 @@ SERVER_NAME="default"
 TYPE="base"
 LAYER="2-services"
 
-COMPONENT_NAME="my-module"
+COMPONENT_NAME="ibm-mq"
 
 mkdir -p .testrepo
 
@@ -51,22 +52,36 @@ else
   sleep 30
 fi
 
-#
-#DEPLOYMENT="${COMPONENT_NAME}-${BRANCH}"
-#count=0
-#until kubectl get deployment "${DEPLOYMENT}" -n "${NAMESPACE}" || [[ $count -eq 20 ]]; do
-#  echo "Waiting for deployment/${DEPLOYMENT} in ${NAMESPACE}"
-#  count=$((count + 1))
-#  sleep 15
-#done
-#
-#if [[ $count -eq 20 ]]; then
-#  echo "Timed out waiting for deployment/${DEPLOYMENT} in ${NAMESPACE}"
-#  kubectl get all -n "${NAMESPACE}"
-#  exit 1
-#fi
-#
-#kubectl rollout status "deployment/${DEPLOYMENT}" -n "${NAMESPACE}" || exit 1
+
+SUBSCRIPTION="subscription/${COMPONENT_NAME}"
+count=0
+until kubectl get "${SUBSCRIPTION}" -n "${NAMESPACE}" || [[ $count -eq 20 ]]; do
+  echo "Waiting for ${SUBSCRIPTION} in ${NAMESPACE}"
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for ${SUBSCRIPTION} in ${NAMESPACE}"
+  kubectl get subscription -n "${NAMESPACE}"
+  exit 1
+fi
+
+count=0
+until kubectl get csv -n "${NAMESPACE}" -o json | "${BIN_DIR}/jq" -r '.items[] | .metadata.name' | grep -q ibm-mq || [[ $count -eq 20 ]]; do
+  echo "Waiting for ibm-mq csv in ${NAMESPACE}"
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for ibm-mq csv in ${NAMESPACE}"
+  kubectl get csv -n "${NAMESPACE}"
+  exit 1
+fi
+
+CSV=$(kubectl get csv -n "${NAMESPACE}" -o json | "${BIN_DIR}/jq" -r '.items[] | .metadata.name' | grep ibm-mq)
+echo "Found csv ${CSV} in ${NAMESPACE}"
 
 cd ..
 rm -rf .testrepo
